Organize functions in logic/config
