--- packages/sanes-chrome-extension/src/logic/config/connection.ts
@@ -1,44 +1,0 @@
-import { bnsConnector } from '@iov/bns';
-import { ethereumConnector } from '@iov/ethereum';
-import { liskConnector } from '@iov/lisk';
-import { ChainConnector } from '@iov/bcp';
-
-import { CodecType } from './configurationfile';
-
-export enum Codec {
-  Bns,
-  Lisk,
-  Ethereum,
-}
-
-export function codecFromString(input: CodecType): Codec {
-  switch (input) {
-    case 'bns':
-    case 'bov':
-      return Codec.Bns;
-    case 'lsk':
-      return Codec.Lisk;
-    case 'eth':
-      return Codec.Ethereum;
-    default:
-      throw new Error(`Codec '${input}' not supported`);
-  }
-}
-
-export function chainConnector(
-  codec: Codec,
-  nodes: ReadonlyArray<string>
-): ChainConnector {
-  const url = nodes[0];
-  switch (codec) {
-    case Codec.Bns:
-      return bnsConnector(url);
-    case Codec.Lisk:
-      return liskConnector(url);
-    case Codec.Ethereum:
-      const scraperApiUrl = nodes[1];
-      return ethereumConnector(url, { scraperApiUrl });
-    default:
-      throw new Error('No connector for this codec found');
-  }
-}
